SucursalDTO

+++ src/main/java/model/Sucursal.java
@@ -1,15 +1,15 @@
 package model;
 
 public class Sucursal {
-    private Integer numero;
+    private Long numero;
     private String direccion;
     private Usuario responsableTecnico;
 
-    public Integer getNumero() {
+    public Long getNumero() {
         return numero;
     }
 
-    public void setNumero(Integer numero) {
+    public void setNumero(Long numero) {
         this.numero = numero;
     }
 
@@ -30,7 +30,7 @@ public class Sucursal {
     }
 
     public void actualizarDatos(
-            Integer numero,
+            Long numero,
             String direccion,
             Usuario responsableTecnico
     ){
